requirements.txt
# LangChain and Ollama
langchain>=0.1.0
langchain-core>=0.1.0
langchain-ollama>=0.1.0
langchain-community>=0.1.0

# MCP SDK
mcp>=0.9.0

# Core dependencies
pydantic>=2.0.0
asyncio

# Vector Store & Embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
chromadb>=0.4.0

# Document Processing
pypdf>=3.17.0
pdfplumber>=0.10.0
python-docx>=1.0.0
python-pptx>=0.6.0
markdown>=3.5.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Image Generation & Processing
diffusers>=0.25.0
transformers>=4.36.0
torch>=2.0.0
torchvision>=0.15.0
accelerate>=0.25.0
pillow>=10.0.0
opencv-python>=4.8.0

# Data Analysis
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0

# Code Execution & Analysis
jupyter>=1.0.0
ipython>=8.12.0
black>=23.0.0
autopep8>=2.0.0
pylint>=3.0.0

# Web & Scraping
requests>=2.31.0
aiohttp>=3.9.0
playwright>=1.40.0
scrapy>=2.11.0

# Audio Processing
openai-whisper>=20231117
pydub>=0.25.0
soundfile>=0.12.0

# Database
sqlite3
sqlalchemy>=2.0.0
chromadb>=0.4.0

# File Processing
chardet>=5.2.0
magic>=0.4.0
python-magic>=0.4.27

# Git Operations
gitpython>=3.1.0

# Utilities
tqdm>=4.66.0
rich>=13.7.0
python-dotenv>=1.0.0
watchdog>=3.0.0
